+++ src/core/proto/payment_http_client.h
@@ -44,7 +44,7 @@ typedef struct utility_response{
 } utility_response_t;
 
 char* tp_create_payment_info(char *url, create_payment_info_t* request); //1
-void tp_http_payment(char *url, process_payment_request_t* request, int hup_num); //2
+void tp_http_payment(char *url, process_payment_request_t* request, size_t hup_num); //2
 void tp_http_command(char *url, utility_command_t* request); //4
 void tp_http_response(char *url, utility_response_t* request); //5
 void ship_log(const char * prefix, const char *url, const char* requestBody, const char* responseBody);
